requirements.txt
requests
python-dotenv
rich


